requirements.txt
asyncio
semantic_kernel
azure.identity
